#85 optimize code
